Move the MemoizeCache to core.utils
