+++ core/utils.py
@@ -1,5 +1,7 @@
-from config.config import GLOBAL_ADMIN_IDS, SERVER_ADMIN_IDS, CARDS
+from discord import Message, TextChannel, User
 from random import choice
+
+from config.config import GLOBAL_ADMIN_IDS, SERVER_ADMIN_IDS, CARDS, MAX_CACHE_LENGTH
 
 
 def trim_quotes(string):
@@ -52,3 +54,44 @@ def draw_random_card_sets(width, groups):
         out += ''.join(cards) + ('\n' if i < groups else '')
 
     return out
+
+
+class MemoizeCache():
+    def __init__(self, bot) -> None:
+        self.bot = bot
+        self.cachedChannels = {}
+        self.cachedMessages = {}
+        self.cachedUsers = {}
+
+    async def get_channel(self, id: int) -> TextChannel:
+        if id not in self.cachedChannels:
+
+            # "Clear the cache" if it gets too big
+            if len(self.cachedChannels) >= MAX_CACHE_LENGTH:
+                self.cachedChannels = {}
+
+            self.cachedChannels[id] = await self.bot.fetch_channel(id)
+
+        return self.cachedChannels[id]
+
+    async def get_message(self, channel: TextChannel, id: int) -> Message:
+        if id not in self.cachedMessages:
+
+            # "Clear the cache" if it gets too big
+            if len(self.cachedMessages) >= MAX_CACHE_LENGTH:
+                self.cachedMessages = {}
+
+            self.cachedMessages[id] = await channel.fetch_message(id)
+
+        return self.cachedMessages[id]
+
+    async def get_user(self, id: int) -> User:
+        if id not in self.cachedUsers:
+
+            # "Clear the cache" if it gets too big
+            if len(self.cachedUsers) >= MAX_CACHE_LENGTH:
+                self.cachedUsers = {}
+
+            self.cachedUsers[id] = await self.bot.fetch_user(id)
+
+        return self.cachedUsers[id]
